Implicit naaarrowing
Address instances of "Implicit narrowing conversion in compound assignment"

+++ src/net/sourceforge/kolmafia/session/BuffBotManager.java
@@ -810,7 +810,7 @@ public abstract class BuffBotManager
 
 		public void updateFreeState()
 		{
-			int totalCost = 0;
+			long totalCost = 0;
 
 			for ( int i = 0; i < this.casts.length; ++i )
 			{
